Propagate declaration to interface for proper optimization of MODULE.Foo(x)
When we call MODULE.Foo(x), we really reference the function
interface, not implementation. Since code generation for record
references depends on whether Foo is a global (in which case we do not
emit a field reference) or a field (in which case we do dereference
the field), we need to propagate field and global flags to the
interface when we have found a matching implementation.