Rank overloads and written forms using renames
This was supposed to eliminate two of the persistent failures,
instantiating 'Pair' and seeing ambiguities in the constructors,
by making the constructor Pair(X : name) be higher precedence than
the constructor Pair (X : Pair). For some reason, however, that
did not work. Still, since there is no regression, I decided
to keep the modification.