Don't lookup in parents when looking up in using-table with a scope
In 'instantiation_of_complex', the reference to IO really refers to an
enclosing scope of one of the 'use' clauses (it's an import of the
used module). This is incorrect, and internally it won't work because
we get a scoped notation as a result of the lookup
(e.g. console_859.TEXT_IO) which we can't resolve in REC.Index because
a record lookup is local only. So the initial lookup must be
consistent with that.

Unfortunately, that doesn't improve the error message in the above
scenario, for a reason that still eludes me.