Don't bother indexing globals found in 'using' lookup
When we have IO.Write, IO might be found in a 'using' lookup, in which case
LookupOne will return something like 'console_589.TEXT_IO'. This is not a real
record index, so we don't want to call REC.Index in that case, since we
know what the machine name is already.