Don't block instantiation just because of recursive validation
With any lookup, we now have written the instantiation of complex in a
way that requires recursive validation of 'writable'. This should not
cause that test to fail.

Unfortunately, something along the way has disabled constructors for
that scenario, so the complex test has a severe case of output
mismatch.