Make sure that module machine names have the extended module type, not 'module'
EnterDeclaration is assigning a type to the machine name in the
general case. However, in the case of modules, this is incorrect, as
we may only know the complete type after we have reconciled /
autoexported the module. A side effect was that 'IsFunction' would
fail because the module record type was not correct (based on machine name),
so we were not finding the module symbols and could not locate the
function.

git-svn-id: https://xlr.svn.sourceforge.net/svnroot/xlr/trunk@532 ed21b5e9-bf16-0410-b9ea-f1dd0686f206