Rank overloads and written forms using renames
This was supposed to eliminate two of the persistent failures,
instantiating 'Pair' and seeing ambiguities in the constructors,
by making the constructor Pair(X : name) be higher precedence than
the constructor Pair (X : Pair). For some reason, however, that
did not work. Still, since there is no regression, I decided
to keep the modification.

git-svn-id: https://xlr.svn.sourceforge.net/svnroot/xlr/trunk@539 ed21b5e9-bf16-0410-b9ea-f1dd0686f206