Don't bother indexing globals found in 'using' lookup
When we have IO.Write, IO might be found in a 'using' lookup, in which case
LookupOne will return something like 'console_589.TEXT_IO'. This is not a real
record index, so we don't want to call REC.Index in that case, since we
know what the machine name is already.

git-svn-id: https://xlr.svn.sourceforge.net/svnroot/xlr/trunk@548 ed21b5e9-bf16-0410-b9ea-f1dd0686f206