Added ClearAttempt, clear all results of validation (incl. locals & globals)
When we were validating a function, we used to generate a number of
temporaries as part of the code generation for the validation. Some of
these would go to the function_context, and remain after validation,
even if they are useless and possibly harmful in the generated code
(e.g. it may be incomplete types, which break the compilation)

git-svn-id: https://xlr.svn.sourceforge.net/svnroot/xlr/trunk@566 ed21b5e9-bf16-0410-b9ea-f1dd0686f206