Baby step #4: Passing parameters in alphabetical order? Bad idea!
It turns out we were sorting parameters based on the parameter map, so
we were sorting them in alphabetical order. This is not exactly how
the code for the built-ins was expecting them. I was puzzled, because
it worked for some code that happened to have "x,y,z" as parameter names :-)

git-svn-id: https://xlr.svn.sourceforge.net/svnroot/xlr/trunk@708 ed21b5e9-bf16-0410-b9ea-f1dd0686f206