A slightly saner model for types
The old model was that a type function was taking a non-type as an argument,
which broke the generally accepted model for generated code and callbacks.

The new model instead generates a prefix rewrite, so that the type check
callback is called on a normal form that looks like 'integer 0'. This
form returns NULL or a type-safe value.