Fixed the 02.Data/rects.xl test
By default, the LLVM JIT is eager and tries to compile functions referenced
by the current one. In our case, we are building the evaluation for one node,
which then builds the evaluation for another one, which refers to the first
one. Since the first one is incomplete, the JIT asserts.

Solution: disable lazy JIT-ing.

Conflicts:

	xl2/xlr/compiler.cpp