LCS algorithm updates
- LCS<T>::Extract now uses an iterator and a call to the T::insert method
to copy the LCS sequence, rather than T::operator+. This allows more
flexibility (for instance, T can now be std::vector<>).

- Add LCS<T>::Extract2 method to extract two copies of the LCS.
One is built by copying elements form the first sequence passed to
LCS<T>::Compute, the other one is built by copying from the second sequence.
Of course this is totally useless if T is e.g., std::string, because in this case
the 2 copies would be identical strings. Things get different when T is a custom
type with a mildly weird operator==.

- A couple of cosmetic fixes.