Make the overload resolution work regardless of definition order
A problem happened if we had the shortest tree first, e.g.:
   truc x:real         -> 1
   truc x:real, y:real -> 2

In that case, if we tried to reduce 'truc 2.2, 3.3', we would first
evaluate truc with a single argument. That would cause us to ultimately
assign storage for 2.2 (as part of an infix tree construction).
When we ran the second case, we'd see the storage as being there, but
it was not necessarily initialized for this form. So we would end up
picking up some uninitialized value (typically, some other tree) and
send that for example to xl_type_check.

The fix is to initialize the tree every time we call NeedStorage.
This is slightly wasteful of IR instructions, but it should get
optimized away.