Added support for partial functions
The problem was that we would stop finding which functions apply when the
argument tree was larger than the name. For instance, say we had

   Foo x,y,z -> ...
   Bar a     -> Foo a,2

we would not call Bar 0,1 and get Foo 0,1,2 as expected. We'd block the matc
of a with x,y, saying that we could not match a single name to a tree.