Generate code for closures with no arguments, fixes #852
Since xl_new_closure returns the original tree when ntrees is 0,
we need to have that returned tree have the 'code' field set by
ArgumentMatch::CompileClosure, because the user of the tree
(e.g. 'shape' in bug #852 source) will receive that original tree
and try to evaluate it.

This can be done safely by compiling "normally" whenever there's no
parameter being passed to the code.