Fix to pass optimized the min-overload test.
We had a problem that we create a tree constant for the 'no match' case
of 'min', which was then used as an initializer by NeedStorage.
However, due to type inference, the expected machine type was 'double', and
we cannot store a tree like 'if x < y then x else y' in a double.

The solution is to avoid initializing from a constant if the constant type
is not correct for the target.