Remove import and load from basics.tbl, closes #1652.
The 'import' and 'load' primitives are special. They need to be processed
at declaration time, not just execution time. Some recent changes,
probably related to #1610, caused this to misbehave on the master branch.

In reality, it's a rather bad idea to have 'import' and 'load' in the
rewrites at all. So I'm removing them. This works on XLR and on the Tao side.