Clone cached arguments in CompileCall, fixes #1981
Consider what happens if we have 'foo 1, 2, 3' as the first call.
We generate code for 'foo 1,2,3' with signature 'foo:III'.
Subsequent calls with the same III signature will simply patch 1, 2 and 3
with the desired value. This may match the original 1, 2 or 3 if we kept it.

Since we now archive the input data in load_data, we have such case. With
data such as:

1,2,3
4,5,6
7,8,9
1,2,3.0
4,5,6.0
7,8,9.1

The second run will show:

7,8,9
4,5,6
7,8,9
7,8,9.1
4,5,6.0
7,8.9.1

In order to avoid that, we need to keep separate copies of the input arguments.