Avoid calling functions with incorrect argument count, fixes #2051
When generating an XLCall with a non-zero arity, the Symbols::calls[]
associative array would contain a tree such as

    error "Couldn't compile $1", 'scrogneugneu'

The 'code' field for that tree used to store the geneated code,
but that was not an actual eval_fn, since it had parameters
(in the above example, two parameters).

So if that tree was later re-evaluated, xl_evaluate() would invoke
that 'code' field with zero arguments, leaving parameters
uninitialized.