Improved fix for #2051, fixes #2077
The original fix for #2051 caused us to not track XLCall trees. So in
conditions where the garbage collector kicks in, we could end up
trying to call code that was pointed to by a tree that no longer
existed.

The alternate fix implemented here returns to the original strategy of
storing a tree, but improves it to keep #2051 fixed.

- We build the call tree, e.g.
     error "Foo", 'bar'
- We return that tree, but its code is not set. That way, if it ever
  executed later, we don't die, we simply compile with fixed args.
- We wrap that tree in a Block shield, and that's where we put the 'code'.
- We don't return the Block, but we store it in the calls map.

The last 3 steps are only performed if there are arguments. Otherwise,
the code pointer remains valid, so we can simply return the original
tree.

Unfortunately, this doesn't fix #2078 or #2076.