Make it so that lookup ordering does not depend on lookup order..., fixes #2660
With #2657, the new 'override_priority' keyword was added, which makes it
possible to change the order in which symbol tables are looked up.

Unfortunately, this keyword was effective only when executed, which
means that its effect actually dependent on the order in which we
loaded files. Unfortunate.

Made this keyword a declarator, but ran into another issue, which is
that if it was declared before imports (as in the case in
slides.xl), then the imports would be looked up in the wrong directory
because the module_dir lookup would place the "downgraded" slides.xl
behind the other modules that declared a module_dir.

So I added an ugly hack, which is that we don't sort symbol tables if
the priority of the current symbol table is above 1e6. This makes it
possible to temporarily disable symbol table ordering.