Fix the priority of 'with' to be an expression, fixes #3043
The issue became visible in expressions such as 'foo with -10..10'.
The change for #3006 puts -10 in 'statement' mode, so we don't parse
expression -10..10 correctly (it parses as -(10..10)).

In order to avoid that, we need 'with' to be part or an expression,
which it is (it is not intended for use in statements).