Improve 'for' loops performance, closes #3303
The solution implemented here is to consider the loop body as a
function with the loop variable as an argument. This means that the
implementation no longer needs to write the symbol table to update
the loop variable, and that the implementation of the loop can refer
to the variable directly as an argument of the callback function.