An additional optimization that works around #3327, fixes #3327
The problem appears with expression
    exp(1) * time

The problem goes away without the parentheses, although I see nothing wrong
with the code generated with the parentheses, just that it's a bit less
efficient because it does an additional type check that is useless.
That being said, the less efficient code causes LLVM to call into an 
unreachable block (which in itself is rather strange).