Evaluate blocks without copying possibly null pointers, fixes #3327
We were starting the evaluation of blocks by copying the LLVM value
associated to the block into its child. But often, the child is where the
evaluation happens, and the block itself has been NULL-initialized.
We may end up with a NULL-deref. In the case of #3327, the optimizer was
smart enough to figure out that null-deref statically, so it replaced it
with a bizarre llvm.trap.

Lesson learned: seeing llvm.trap in the code means LLVM detected a NULL-deref.

Also removed the now useless block evaluation routines from basics.tbl