Don't index array if an A[B] form exists, closes ticket #10
The investigation of this bug also showed that the variable used to
count stack overflows was not xl_recursion_count but a variable by the same
name created by LLVM. This is annoying for debugging, or if we later want
to use xl_recursion_count in C++ code. This is fixed.