Massive re-engineering of the TBL files, refs #3555 @8h
The main objective of this commit is to make it simpler to create TBL files
in an application that uses XL. With this commit, you simply do:

    #include "opcodes.h"
    #include "my-extension.tbl"

and everything happens for you in the background. Specifically, the macros
in opcodes.h will register all names, types, functions, etc, and then
enter them automatically in the symbol table at startup time. This is done
using C++ initialization, which is a bit unreliable regarding ordering. So
the mechanisms implemented here first do a simple registration, and do the
actual work in main, once we known that everything has been registered.