Change the syntax from 'X -> Y'  to 'X is Y' for readability
Historically, the -> operator was seen as a "transforms into", but
it's really hard to read for people not familiar with the language. So
it was changed to 'is', which conveys the meaning much better for newbies.

Compare new:                            old:
  X is 0                                X -> 0

  0! is 1                               0! -> 1
  N! is N * (N-1)!                      N! -> N * (N-1)!

  sin X is cos(pi/2+X)                  sin X -> cos(pi/2+X)

No test were harmed in the process, but you obviously need to udpate
your source code...