Run tests locally in build directory instead of after installation
This is safer, as it makes it possible to run tests before installing in
the target installation directory.

The notion of "search path", which up to now was delegated entirely to
the application, and had so far only really been implemented in Tao3D
using the Qt notion of path search, has been pulled back into XL.

Next step will be to add options to test in the installation directory.