Remove garbage-collected compiler data entirely
The problem with keeping the data in a tree info node is that it
is not correctly dependent on the function or context we are in.
So it's better to keep it in a local map. This also makes gets rid of
the need for all the somewhat uncomfortable garbage-related deletion
of LLVM data.

In order to do that, the capture information for closures also had to
be moved explicitly to types. Fortunately, its' not too hard to
compute. The capture generation code has been reworked, but is
probably not functional yet.