Compute the returned type for declarations
This fixes the following case:

    3 mod 5

when using the definition in builtins.xl. This is because the function
is a C function, and its type is deduced from the declaration [as]
clause. Without evaluation, we get a new [integer] name, not the
original one, so autoboxing was not correctly generating the target
type.