Add special primitive IAbs for integer absolute value
For some computation kinds, the code sequence is the same irrespsective
of the input type. Calling a runtime support function for this is
problematic, since this will be inefficient for sized data types,
e.g. i32. So in order to benefit from the polymorphic aspects of LLVM,
we introduce 'special' opcodes that generate the required code
sequence.

There are many candidates functions for this approach that will be
introduced incrementally.