Add a template Arg to Error class to avoid conflicts on 32-bit machines
The problem is to have Arg(long long) and Arg(double). That causes
an ambigutity with Arg(int), so overloads were added for other types,
but depending on machine, long and long long are the same type or
different types. The way it was previously written caused errors on
32-bit machines.

The new approach uses a template that will be instantiated exactly as
needed.