Fix regression for closures
The problem was introduced by the new, better way to generate constants tree pointers.
The code no longer correctly detected if a tree was an input argument to the function
while initializing the result, and generated code overwriting the input argument
with a constant pointer right after reading the argument:

  store %Tree* %1, %Tree** %result
  store %Tree* inttoptr (i64 87747720 to %Tree*), %Tree** %result

The fix is to set value[src] for input arguments.