Correctly link when LLVM is compiled without RTTI
When LLVM is compiled without RTTI, we end up with a run-time error
for missing RTTI symbols, e.g. _ZTIN4llvm3orc14SymbolResolverE
(typeinfo for llvm::orc::SymbolResolver).

We don't really use RTTI in LLVM, only in the XL front-end, mostly
to generate debug messages or recorder entries. So it's OK to disable
it while compiling llvm-crap.cpp, but that means we have to skip
some GC header portions that refer to typeid().