Some fixes to LLVM7 builds
I had made some bad decisions based on a custom-build of a pre-release
of LLVM7, where some decisions about reverting back to std::unique_ptr
had not solidified yet.

Lesson learned: trying to follow pre-release LLVM variants is OK to be
aware of what's coming, but there's a risk that this embeds incorrect
decisions until the release is actually published.