Add ability to FILTER the results of a command
Some results are not portable. For example, symbol names
may be prefixed or not by a _ in the error message output,
or the result of puts may be a non-portable value.

Running a filter command makes it possible to normalize
the output in order to get consistent result across platforms