Evaluate matching infix name in value types, not binding types
Consider the case where we call

   write Items

from

   write_line Items

with argument [1, 2, 3]. In that case, Items is a variable with a type
that can be known at runtime to be an infix, but we just get a pointer
to the infix. The runtime then calls special form [name Items] to
extract the name of the infix. That needs to be evaluated as part of
argument values (here [Items]), not as part of bindings, since we
don't bind anything to [Items].