Fix issue with DYLD_LIBRARY_PATH not working on Catalina
In Catalina with System Integrity Protection (SIP) enabled,
the DYLD_LIBRARY_PATH is simply ignored. According to the manual,
this is only for binaries protected by SIP, but apparently, this
is also true for binaries you build, and even if no library
would be found without DYLD_LIBRARY_PATH. I also tried
the DYLD_FALLBACK_LIBRARY_PATH, with the same behavior.

For now, simply link the files in place. A better fix will
be forthcoming in make-it-quick