symbols: Switch to "larger and more specialized" ordering
The symbol table structure is completely changed to accomodate an ordering of
patterns according to the "larger and more specialized wins" ordering.

This means that now, the following programs are equivalent:

     0! is 1
     N! is N * (N-1)!

and

     N! is N * (N-1)!
     0! is 1

In both cases, 0! will win over N! because it is more specialized.

This requires a deep reorganization of the symbol table, which now looks closer
to what the actual in-code representation looks like, but with an implicit
ordering of entries. See the comments in context.h for detail.

A side effect of this refactoring is that the Rewrite and Scope classes now have
a real C++ distinct C++ class, as opposed to being typedefs. This makes it
possible to have template specializations for Tree::As<Scope>() and
Tree::As<Rewrite>() that actually check that you are doing something sane.
This also will help catch structure errors in the symbol table.

The symbol table now contrains Scopes(), which are a prefix of the enclosing
scope and the current scope. Each scope is of class Scope, which is a block
containing either nil (empty scope), or a Rewrite (i.e. a form like
[Pattern is Implementation]), or Rewrites, i.e. a form like [Left;Right], where
[Left] is a Rewrite, and [Right] is either a Rewrite or a Rewrites.
If it's a Rewrites, then the tree is sorted so that the current [Left] is
between [Right.Left] and [Right.Right].

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>