types: Rename the Types class to CompilerTypes
This is a step towards separating the semantics parts of the Types class, which
presumably shold be reusable by the interpreter, and the compiler-specific
aspects of it.

The intent is to soon add a Types class that does new-style semantics analysis,
and then connect whatever can be salvaged from CompilerTypes in it. If anything.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>