types: Split the RewriteXYZ classes into based and compiler
Like the rest of the type analysis, classes like RewriteCandidate and
RewriteBidnings should really be shared between the interpreter and the
coompiler. Split between a base class that does the semantics analysis and a
compiler variant that performs the JIT-related operations and maintain the
JIT-specific data structures.

This is a rather big refactoring that was actually performed over a rather long
period of time with a number of intermediate commits. The intermediate commits
were fused because most of them were causing major regressions in the test
suite, so this would mess up git bisect.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>