xldebug: Automatically switch to "Scope" and "Rewrite" mode
When using xldebug, it's annoying that a Scope is a Prefix and a local symbol
table is an Infix. When you print them through the normal renderer, it comes out
in a way that is really not comfortable.

We can auto-detect based on rather atypical content if this looks like a Scope
or a symbol table, and automatically switch to Context::Dump in that case.

Also minor improvement on the output formatting - Spaces were missing

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>