llvm: Repair -B option
The -B option was dependent on which specific version of LLVM was being used,
since it was placed in module optimization code that does not evaluate on all
versions.

This is now moved to compiler destruction, which exposed the fact that there
were a few bugs in that compiler cleanup code.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>