modules: Update make-it-quick and recorder, fix variables
Variables in make-it-quick changed names, so we need to make some changes in the
xl2 makefiles. There was also a bug in make-it-quick itself which broke the
installation of shared libraries. This is fixed by picking up the new version of
the submodule.

According to changes in the makefile, replace
MAKE_CXX with COMPILE.cpp,
EXE_EXT with EXT.exe
OBJ_EXT with EXT.obj

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>