native: Connect to interpreter
Adding two small test functions for testing purpose.

This looks promising, but the interpreter itself is broken at the moment with
respect to calling getpid.

Also, in the long term, we will probably need to put native functions in
modules, e.g. refer to [getpid] as [POSIX.getpid] or something like that.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>