interpreter: Process scopes correctly
This now accepts the dotted notation to refer to an object,
and the prefix notation to refer to the enclosing scope.

Example:

    thing is
        A is 3
        B is 4

    A is 8

    thing.A + thing.(B) + {X is 3; Y is 7} (X*Y+A)

This computes 36 = 3 + 4 + 3 * 7.

Notice that the following will fail, because there is not (local) form matching
[A+B] inside [thing].

    thing is
        A is 3
        B is 4
    A is 8
    thing.(A + B) + {X is 3; Y is 7} (X*Y+A)

If there is a local definition, then it works, and the code below computes 26:
    thing is
        A is 3
        B is 4
        X+Y is -3
    A is 8
    thing.(A + B) + {X is 3; Y is 7} (X*Y+A)

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>