types: The value types must be consistent across candidates
In the Fibonacci example, you want the type of [fib(N-1)] to be the same whether
you are calling [fib 0] or [fib 1] or [fib N]. That implies that the type must
be kept in the values context, not in the binding context.

This is a better solution to have the Factorial and Fibonnacci tests pass.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>