compiler: Cache the value when there is a chance we will reuse it
While investigating the `library-writeln-list` test, it appears that the
generated code is bad because we call `xl_write_text` with a `Tree *` instead of
a `char *`. This exposes that there is another problem: the kind checks cannot
correctly distinguish between `character` and `text` since both have the same
`TEXT` kind. So type checking needs to be re-done.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>