language: Update basic tree to use natural instead of integer
This allows for larger values in the tree representation.
The value -1 is an integer, but value 1 is a natural.

Temporarily, we print natural values as signed values, because the type system
is still not strong enough to "do the right thing", so this limits the number of
regressions in the test suite.

Despite this, there are a few hard to explain regressions in the LLVM test
suite, notably builtin-mod. They have been marked by updating the baseline for
now, since the objective is to make progress on the interpreter for the moment.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>