modules: Move module handling to Module class
All the old "SourceFile" crap (and really ugly code) has been cleaned up in
favor of the Module class. This results in better error messages, but also adds
new options, like the `-preload` option to add more files to pre-load.

This also fixes a nagging bug in the old compiler, where evaluating files `a.xl`
then `b.xl` would evaluate `b.xl` within the scope of `a.xl`.

The result is printed only if not `nil`. This makes it easier to create programs
that don't show anything.

A side effect of the new module processing is that you can now pass the files to
compile without the extension, e.g. `./xl a` will compile `a.xl`.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>