builtins: Add implicit conversion from integer to real
Three is no `IAbs` / `FAbs`. It is much better to define `abs` in `builtins.xl`,
but this requires `X < 0` to also work for `real` values.

As a first step, introduce implicit conversion from integer to real.
This does not work yet, because the call of implicit conversions is not there
yet. This requires additional logic to make sure that an implicit conversion
does not pre-empt a direct value.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>