native: Fix a bug recently introduce in the indexing of arguments
This was discovered while writing a reduced case about the technique for a blog
entry. Using a simple `index++` generates a warning about sequence point. This
was seen in the original code (which used `args++`), and was the reason for
switching to a mistaken use of `sizeof...`, which happened to work because it
was only ever used so far with a single extra argument.

Using a wrapper around the index increment probably does not fix the actual
sequence point issue, but testing shows that the compiler for now at least does
not attempt to reorder calls in a variadic parameter pack expansion. We can hope
that this remains true in the future.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>