interpreter: Perform evaluation of bodies with STATEMENT
The case of complex division requires a body that has local declarations.
We need to process these declarations in order to evaluate properly.

Then we need to strip any local closure from the returned value in order to be
able to do the pattern matching e.g. with an infix.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>