modules: Do not pass a scope when creating a module
Modules are created in the top-level scope, so there is no need to pass a scope
to Modules::Get(). In addition, passing `nullptr` to identify module import
during early parsing was fragile, and could cause the creation of malformed
symbol tables.

In addition, a bit of further cleanup on the interfaces for the module
functions, e.g. returning `Tree *` instead of `Tree_p` when the value is known
to be owned by the module anyway.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>