modules: Lookup for files following XL capitalization rules
An exact match is preferred. Depending on the setting of the `case_sensitive`
option, this is either a file matching the exact spelling of the source code, or
an all-lowercase, no-underscore spelling. This makes it possible to have
slightly different module implementations based on the capitalization option,
which may not be a very good idea, frankly.

If this exact match is not found, then the file is searched with the expected
extension, but ignoring the capitalization of the rest of the file name. For
example, looking for `MY_SUPER_MODULE` would find `mysupermodule.xs`,
`MySuperModule.xs` or `MY_SUPER_moDULE.xl`. Whether `mysupermodule.XS` will be
foudn depends on whether the filesystem is case-sensitive or not. The rationale
for treating the extension differently is that it may have a meaning to the
operating system. For example, in the early days of C++, some people thought
that using the `.C` extension for C++ was a bright idea.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>