modules: Reverse priority for qualified statements like IO.print
For example, if we have [IO.print "Hello"] as a statement, we need to
rewrite it as [IO.{print "Hello"}] in order to lookup the prefix in the correct
context.

For other notations such as infix or postfix, it is assumed that you need the
explicit parentheses, e.g. [IO.(A + B)] or [IO.(A%}]. This is both clearer and
easier to implement since there is essentially nothing to do.

This change also introduces the distinction between sequences and statements
during evaluation, which is likely to be useful if only for debugging.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>