math: Implementation of the math module using more modern syntax
The math module is one of the easiest modules to convert to the "new" syntax for
module specifications.

Functions there are lifted from `<math.h>` on Linux. Presumably, the functions
are generally available on all platforms. In the long run, it would be
interesting to write all of them in XL, but for now it's mostly a waste of time.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>