context: Make it possible to match more than one prefix/postfix
This makes it possible to match patterns such as `sin sin X`.

This also exhibited an existing bug in 04-basic-operators-in-function.xl,
which is that `-5 rem 2` was not matching `X:natural rem Y:natural` for an
entirely wrong reason, which was that the sorting was incorrect because of the
prefix `-` in front of `5`:

    [1658258 1.398755] symbols_sort:       SEARCH 0x7fd38b161620 [X:natural rem Y:natural] < 0x7fd38b14b300 [- 5 rem 3]

In other words, it was a bug in the sorting for the symbol table that excluded
the version of `rem` for `natural`, not a type check. Further testing showed
that `-5 as natural` passes, so this is really a pre-existing bug and not a
regression. Updating the test result for now, and adding a test to the baseline
that will need fixing.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>