interpreter: Flat evaluator (While loop optimization)
Refactor the main interpreter functions to avoid deep recursion.
The interpreter now simply lookups the prototype, exits at first prototype that
matches, and switches contexts to evaluate the body of the prototype in the
bound scope. As a result, it never goes into any deep recursion when evaluating
a tail-recursive body like the standard definition of the while loop.

Also a small optimization in Context::Enclose to avoid creating closures of
closures when passing parameter around. This causes an indirect memory leak,
where each recursion would pass a closure with `Body` as its expression,
implicitly keeping a stack of all the caller frames that are no longer useful.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>