bytecode: Add a basic bytecode interpreter
At this stage, the bytecode interpreter merely compiles, and does not
interfere with the behavior of the base interpreter.

This is a squash commit of work done in a separate branch on top of the
interpreter branch, and rebased regularly, as some of the improvements made
sense for the base interpreter branch as well.

The purpose of the bytecode interpreter is to be as flexible in terms of
evaluation as the base interpreter (i.e. capable of evaluating generated code,
etc), while at the same time being significantly faster. To do that, the work
began by defining what the primary evaluation loop would look like, and to make
it so that it would be as streamlined as possible.

In the incarnation proposed here, the evaluation loop in Bytecode::Run looks
exactly like this:

    Bytecode *bc = this;
    while (bc)
    {
        opaddr_t max = bc->code.size();
        state.pc = 0;
        state.bytecode = bc;
        while (state.pc < max)
            bc->code[state.pc++](state);
        bc = state.bytecode;
    }

In other words, evaluating opcodes ia tight loop doing one indirect call per
opcode, the indirect code passing a single pointer argument to a RunState.
I don't think I can do much tighter in C++.

Opcodes are defined by functions with the following prototype (aka opcode_fn):

    void opcode_implementation(RunState &state)
    {
        ...
    }

The RunState contains an evaluation stack, a pointer to the current scope, a
pointer to the currently executing bytecode, a program counter in that bytecode,
the number of locals / input arguments (which are put at the beginning of the
stack), and an error pointer for error reporting.

This puts some constraints on the bytecodes, which are not objects, but function
pointers. As a result, they cannot carry data. When data is needed, it is stored
as indexes in the bytecode array, which an opcode can access from the state
by fetching it from the current program counter. Depending on the opcode, the
index can indicate:

- A stack location, e.g. for the `local` opcode. This is the reason why local
  temporaries are stored at the beginning of the stack.

- A constant tree, which is stored in the Bytecode object

- A definition Rewrite pointer, also stored in the Bytecode object

- A target program counter, which is used for conditional or unconditional
  branches.

Such branches are generated with a 0 value, and patched after the fact. There
are presently two kinds of branches, on evaluation mismatch, or on evaluation
success.

The rest of the implementation follows from this basic idea of keeping the
evaluation loop as tight and simple as possible.

The code as checked in here is entirely untested. It is merely an attempt at
replicaetd a (simplified) version of the interpreter in byte-compile mode.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>