bytecode: Implement simple anonymous functions
This makes the following code work: [(lambda X is X + 1) 31]

The change is relatively major, because it exposed a problem with the
StorageIndex interface, namely that in some cases, it requires a copy from a
constant. In the above case, for example, we now generate the following code:

    Compiled [(lambda X is X + 1) 31] as
    Constants:
      C0	= 31
      C1	= lambda X
    Locals:
      L0	= [(lambda X is X + 1) 31][31]
    Types:
      31	as	natural
    Opcodes (1 locals, 0 parameters, 1 temporaries)
         0: copy                C0, L0
         3: call                C1, (1)
           0: C0
           => L0
         8: ret                 L0

Having to emit a `copy` from constant `C0` to local `L0` in order to be able to
write the evaluated value of `31` (which in that particular case is strangely
_not_ a constant) means that we need to move all `StorageIndex` references just
before an opcode (we can't insert the copy inside arguments, for example).

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>