bytecode: Call the correct constructor for RunValue()
There is no constructor for `RunValue(0.0, real_mtype)`, so we end
up calling the wrong one. For some reason, the compiler calls one with an
integral type as a first argument. This seems like a compiler bug.

Removing the `mtype` argument allows us to use one of the carefully generated
per-type constructors.

For the following test case:

    foo X is quote(X + 2 * [[X+X]])
    print "foo 3=", foo 3, " foo 5.3=", foo 5.3

Before the fix, the result was:

    foo 3=X + 2 * 6 foo 5.3=X + 2 * 4.94066e-323

After the fix, the result is:

    foo 3=X + 2 * 6 foo 5.3=X + 2 * 10.6

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>